backend/pyproject.toml
[project]
name = "mometrics"
version = "0.1.0"
description = "Backend service for MoMetrics"
authors = [
    {name = "Gennadiy Samoylov",email = "samoylowfgg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (==4.0.1)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "isort (>=7.0.0,<8.0.0)",
    "redis (>5.0.2,<=5.2.1)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.3.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.14.5,<0.15.0)",
    "black (>=25.11.0,<26.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
